video upload and cleanup Fixes

+++ meTube-Backend/src/utils/cloudinary.js
@@ -15,15 +15,21 @@ const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null;
         const response = await cloudinary.uploader.upload(localFilePath, { resource_type: 'auto' })
-        console.log("URL created on cloudinary: ", response.url);
+        // console.log("URL created on cloudinary: ", response.url);
         return response;
 
     } catch (error) {
-        console.log("removed local file after failing to create URL");
+        // console.log("removed local file after failing to create URL");
         return sendError(500, "File upload failed");
     } finally {
-        fs.unlinkSync(localFilePath)
-        console.log("removed local file after creation of URL");
+        try {
+            if (localFilePath && fs.existsSync(localFilePath)) {
+                fs.unlinkSync(localFilePath);
+                // console.log("Removed local file after upload.");
+            }
+        } catch (cleanupError) {
+            console.log("Error removing local file:", cleanupError);
+        }
         // to remove the locally saved temp file from multer as upload operation failed
     }
 }
